@@ -2,10 +2,12 @@
 
 module BlueEyes
   module Extensions
+    # BlueEyes String Extensions
     module String
       def camel_case
-        return self.gsub(/^./) { |l| l.capitalize } if !match(/[_-]/)
-        altered_self = self.downcase.capitalize
+        return gsub(/^./, &:capitalize) unless match?(/[_-]/)
+
+        altered_self = downcase.capitalize
         altered_self.scan(/[_-][a-zA-Z]/).each do |match|
           altered_self.gsub!(match, match[1].upcase)
         end
@@ -14,26 +16,26 @@ module BlueEyes
       end
 
       def camel_case!
-        self.replace camel_case
+        replace camel_case
       end
 
       def directory_name
-        self.downcase.gsub(/[^a-z|\-|\_]/, '')
+        downcase.gsub(/[^a-z|\-_]/, '')
       end
 
       def file_name
-        self.gsub(/[\-| ]/, '_').
-             gsub(/([A-Z]+|[A-Z][a-z])/) { |x| "_#{x}" }.
-             sub(/^_/, "").
-             gsub(/_{2,}+/, "_").
-             downcase
+        gsub(/[-| ]/, '_')
+          .gsub(/([A-Z]+|[A-Z][a-z])/) { |x| "_#{x}" }
+          .sub(/^_/, '')
+          .gsub(/_{2,}+/, '_')
+          .downcase
       end
 
       def file_name!
-        self.replace file_name
+        replace file_name
       end
     end
   end
 end
 
-String.send(:include, BlueEyes::Extensions::String)
+String.include BlueEyes::Extensions::String
