@@ -1,7 +1,9 @@
-require "yaml"
+# frozen_string_literal: true
+
+require 'yaml'
 
 # Redis Configuration
 unless ENV['RACK_ENV'] == 'test'
-  redis_settings = YAML::load_file("config/redis.yml")
+  redis_settings = YAML.load_file('config/redis.yml')
   REDIS = Redis.new(redis_settings[ENV['RACK_ENV']])
 end
