@@ -1,13 +1,17 @@
-require_relative "spec_helper"
+# frozen_string_literal: true
 
-def app
-  ApplicationController
-end
+require_relative 'spec_helper'
 
 describe ApplicationController do
-  it "responds with a welcome message" do
+  before do
     get '/'
+  end
+
+  it 'responds with HTTP status 200 OK' do
     expect(last_response.status).to eq(200)
-    expect(last_response.body).to include("Welcome to the Sinatra Template!")
+  end
+
+  it 'responds with a welcome message' do
+    expect(last_response.body).to include('Welcome to the Sinatra Template!')
   end
 end
