@@ -1,18 +1,22 @@
-require "thor/group"
-require "active_support/inflector"
+# frozen_string_literal: true
+
+require 'thor/group'
+require 'active_support/inflector'
 
 module BlueEyes
   module Generators
+    # Generates an ActiveRecord model
     class ModelGenerator < Thor::Group
       include Thor::Actions
+
       attr_reader :file_name, :class_name, :model_name, :migration_name, :migration_class_name, :table_name
 
-      desc "Generate an ActiveRecord model"
-      argument :name, type: :string, desc: "Name of the model"
-      argument :attributes, type: :array, default: [], banner: "field:type field:type"
+      desc 'Generate an ActiveRecord model'
+      argument :name, type: :string, desc: 'Name of the model'
+      argument :attributes, type: :array, default: [], banner: 'field:type field:type'
 
       # --no-migration make migrations optional
-      class_option :migration, type: :boolean, default: true, desc: "Generate migration for model"
+      class_option :migration, type: :boolean, default: true, desc: 'Generate migration for model'
 
       def self.source_root
         File.dirname(__FILE__)
@@ -27,32 +31,37 @@ module BlueEyes
         @migration_class_name = migration_name.camel_case
 
         attributes.map! do |attribute|
-          field = attribute.split(":")
-          { name: field[0], type: field[1] || "string" }
+          field = attribute.split(':')
+          { name: field[0], type: field[1] || 'string' }
         end
       end
 
       def create_model
         unless model_name == name
-          say "[WARNING] The model name '#{name}' was recognized as a plural, using the singular '#{model_name}' instead."
+          say '[WARNING] ' &
+              "The model name '#{name}' was recognized as a plural, using the singular '#{model_name}' instead."
         end
 
-        template "model.rb.erb", File.join("app/models", "#{file_name}.rb")
+        template 'model.rb.erb', File.join('app/models', "#{file_name}.rb")
       end
 
       def create_migration
         return unless options[:migration]
 
-        migration_files = Dir.entries("db/migrate").select { |path| !File.directory? path }
+        migration_files = Dir.entries('db/migrate').reject { |path| File.directory? path }
 
-        if duplicate = migration_files.find { |file| file.include?(migration_name) }
+        if (duplicate = migration_files.find { |file| file.include?(migration_name) })
           say_status :identical, "db/migrate/#{duplicate}", :blue
         else
-          version = Time.now.utc.strftime("%Y%m%d%H%M%S")
+          version = Time.now.utc.strftime('%Y%m%d%H%M%S')
           migration_file_name = "#{version}_#{migration_name}.rb"
 
-          template "migration.rb.erb", File.join("db/migrate", migration_file_name)
+          template 'migration.rb.erb', File.join('db/migrate', migration_file_name)
         end
+      end
+
+      def create_spec
+        template 'model_spec.rb.erb', File.join('spec/models', "#{file_name}_spec.rb")
       end
     end
   end
