@@ -1,14 +1,17 @@
-require "thor/group"
+# frozen_string_literal: true
+
+require 'thor/group'
 require 'active_support/inflector'
 
 module BlueEyes
   module Generators
+    # Generates Rails-style scaffold
     class ScaffoldGenerator < Thor::Group
       include Thor::Actions
 
       desc "Generate an ActiveRecord model with it's associated views and controllers"
-      argument :name, type: :string, desc: "Name of the model"
-      argument :attributes, type: :array, default: [], banner: "field:type field:type"
+      argument :name, type: :string, desc: 'Name of the model'
+      argument :attributes, type: :array, default: [], banner: 'field:type field:type'
 
       def create_model
         ModelGenerator.new([name, attributes]).invoke_all
